move lootrate float to double

+++ invocation-service/src/main/java/imt/production/dev/Model/ResourceMonstre/MonstreResource.java
@@ -16,7 +16,7 @@ public class MonstreResource {
     private int def;
     private int vit;
     private List<Competence> skills;
-    private float lootRate;
+    private double lootRate;
